+++ CHANGELOG.md
@@ -39,6 +39,7 @@
   * Support filtering MRs by whether they are approved or not.
   * Support setting `merge_commit_template` when creating a project.
   * Support setting `mr_default_target_self` when creating a project.
+  * Support IP restriction ranges when editing groups.
 
 ## Changes